fix: Failed to download some files. #9

+++ pkg/pd/response.go
@@ -32,11 +32,11 @@ type ResponseDownload struct {
 type ResponseFileInfo struct {
 	ID                string    `json:"id"`
 	Name              string    `json:"name"`
-	Size              int       `json:"size"`
-	Views             int       `json:"views"`
-	BandwidthUsed     int       `json:"bandwidth_used"`
-	BandwidthUsedPaid int       `json:"bandwidth_used_paid"`
-	Downloads         int       `json:"downloads"`
+	Size              int64     `json:"size"`
+	Views             int64     `json:"views"`
+	BandwidthUsed     int64     `json:"bandwidth_used"`
+	BandwidthUsedPaid int64     `json:"bandwidth_used_paid"`
+	Downloads         int64     `json:"downloads"`
 	DateUpload        time.Time `json:"date_upload"`
 	DateLastView      time.Time `json:"date_last_view"`
 	MimeType          string    `json:"mime_type"`
@@ -68,12 +68,12 @@ type FileGetList struct {
 	Success       bool      `json:"success"`
 	ID            string    `json:"id"`
 	Name          string    `json:"name"`
-	Size          int       `json:"size"`
+	Size          int64     `json:"size"`
 	DateCreated   time.Time `json:"date_created"`
 	DateLastView  time.Time `json:"date_last_view"`
 	MimeType      string    `json:"mime_type"`
-	Views         int       `json:"views"`
-	BandwidthUsed int       `json:"bandwidth_used"`
+	Views         int64     `json:"views"`
+	BandwidthUsed int64     `json:"bandwidth_used"`
 	ThumbnailHref string    `json:"thumbnail_href"`
 }
 
@@ -89,12 +89,12 @@ type ResponseGetUser struct {
 	Username            string              `json:"username"`
 	Email               string              `json:"email"`
 	Subscription        GetUserSubscription `json:"subscription"`
-	StorageSpaceUsed    int                 `json:"storage_space_used"`
+	StorageSpaceUsed    int64               `json:"storage_space_used"`
 	IsAdmin             bool                `json:"is_admin"`
-	BalanceMicroEur     int                 `json:"balance_micro_eur"`
+	BalanceMicroEur     int64               `json:"balance_micro_eur"`
 	HotlinkingEnabled   bool                `json:"hotlinking_enabled"`
-	MonthlyTransferCap  int                 `json:"monthly_transfer_cap"`
-	MonthlyTransferUsed int                 `json:"monthly_transfer_used"`
+	MonthlyTransferCap  int64               `json:"monthly_transfer_cap"`
+	MonthlyTransferUsed int64               `json:"monthly_transfer_used"`
 	FileViewerBranding  interface{}         `json:"file_viewer_branding"`
 	FileEmbedDomains    string              `json:"file_embed_domains"`
 	SkipFileViewer      bool                `json:"skip_file_viewer"`
@@ -106,22 +106,22 @@ type GetUserSubscription struct {
 	Name                string `json:"name"`
 	Type                string `json:"type"`
 	FileSizeLimit       int64  `json:"file_size_limit"`
-	FileExpiryDays      int    `json:"file_expiry_days"`
-	StorageSpace        int    `json:"storage_space"`
-	PricePerTbStorage   int    `json:"price_per_tb_storage"`
-	PricePerTbBandwidth int    `json:"price_per_tb_bandwidth"`
-	MonthlyTransferCap  int    `json:"monthly_transfer_cap"`
+	FileExpiryDays      int64  `json:"file_expiry_days"`
+	StorageSpace        int64  `json:"storage_space"`
+	PricePerTbStorage   int64  `json:"price_per_tb_storage"`
+	PricePerTbBandwidth int64  `json:"price_per_tb_bandwidth"`
+	MonthlyTransferCap  int64  `json:"monthly_transfer_cap"`
 	FileViewerBranding  bool   `json:"file_viewer_branding"`
 }
 
 type FileGetUser struct {
 	ID                  string    `json:"id"`
 	Name                string    `json:"name"`
-	Size                int       `json:"size"`
-	Views               int       `json:"views"`
-	BandwidthUsed       int       `json:"bandwidth_used"`
-	BandwidthUsedPaid   int       `json:"bandwidth_used_paid"`
-	Downloads           int       `json:"downloads"`
+	Size                int64     `json:"size"`
+	Views               int64     `json:"views"`
+	BandwidthUsed       int64     `json:"bandwidth_used"`
+	BandwidthUsedPaid   int64     `json:"bandwidth_used_paid"`
+	Downloads           int64     `json:"downloads"`
 	DateUpload          time.Time `json:"date_upload"`
 	DateLastView        time.Time `json:"date_last_view"`
 	MimeType            string    `json:"mime_type"`
@@ -134,7 +134,7 @@ type FileGetUser struct {
 	CanEdit             bool      `json:"can_edit"`
 	ShowAds             bool      `json:"show_ads"`
 	AllowVideoPlayer    bool      `json:"allow_video_player"`
-	DownloadSpeedLimit  int       `json:"download_speed_limit"`
+	DownloadSpeedLimit  int64     `json:"download_speed_limit"`
 }
 
 type ResponseGetUserFiles struct {
@@ -146,7 +146,7 @@ type ListsGetUser struct {
 	ID          string      `json:"id"`
 	Title       string      `json:"title"`
 	DateCreated time.Time   `json:"date_created"`
-	FileCount   int         `json:"file_count"`
+	FileCount   int64       `json:"file_count"`
 	Files       interface{} `json:"files"`
 	CanEdit     bool        `json:"can_edit"`
 }
